keep tests where they are
